Updated Firmware
Changed over from pulsing LED pattern to combat FM receiver noise on startup.  The LED pattern is now as follows:

~2sec full on @ startup
150 ms blink on button press
500 ms blink on save station

!! There is still a known bug !!
After saving a station, the following button press is not read. It takes two button presses after a save in order to increment a station.  This needs to be fixed.

NOTES:  The corresponding hex file will have this commit ID attached.